requirements.txt
Django==2.2.3
django-rest-auth==0.9.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pillow==6.0.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
django-cors-headers==3.0.2
dj-static==0.0.6 
gunicorn==19.1.1 
